requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
eval_type_backport==0.2.0
fakeredis==2.24.1
filelock==3.16.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
logger==1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ratelimit==2.2.1
redis==5.0.8
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rq==1.16.2
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
together==1.2.12
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
